% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/headlines.R
\name{GetHeadlines}
\alias{GetHeadlines}
\title{Headlines}
\usage{
GetHeadlines(org.file = "~/tmp.org")
}
\arguments{
\item{org.file}{a file path point to a .org file}
}
\value{
a table of headlines and attributes
}
\description{
org headlines
}
\details{
A function to parse org files, will return headlines and associated attributes, including tag, clock entries, shedules, deadlines, closed date, todo states,
}
\author{
Yi Tang
}

