% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/headlines.R
\name{search.parent}
\alias{search.parent}
\title{search.parent}
\usage{
search.parent(head.info, heading.id)
}
\arguments{
\item{head.info}{a head table from GetHeadlings}

\item{heading.id}{a unique id from head.info}
}
\value{
a data.table
}
\description{
Search for parent headlines
}
\details{
Given a headlines table and headline id, it will return the parent headlines.
}
\author{
Yi Tang
}

