% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/headlines.R
\name{GetTS}
\alias{GetTS}
\title{function}
\usage{
GetTS(str = "a", ts.format = c("<\%Y-\%m-\%d \%a>",
  "<\%Y-\%m-\%d \%a \%H:\%M>"))
}
\arguments{
\item{str}{content of a .org file}

\item{ts.format}{format of time stamps used in the .org file. It is equivalent to \code{org-time-stamp-formats} in Emacs}
}
\value{
a
}
\description{
ad descrition
}
\details{
lalla details
}
\author{
yitang
}

