% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clock.table.R
\name{ToISOdate}
\alias{ToISOdate}
\title{clock.table}
\usage{
ToISOdate(clock.entries)
}
\arguments{
\item{clock.entries}{a standard clock entry from org-mode}
}
\value{
POXICt object
}
\description{
Parse clock entry to ISO date
}
\examples{
str <- c("CLOCK: [2014-11-26 Wed 09:36]--[2014-11-26 Wed 10:04] =>  0:28",
         "CLOCK: [2014-12-04 Thu 15:24]--[2014-12-04 Thu 16:25] =>  1:01")
ToISOdate(str)
}
\author{
Yi Tang
}

