% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/headlines.R
\name{tree.headings}
\alias{tree.headings}
\title{org-headings-tree}
\usage{
tree.headings(head.info, output = "screen", plantuml = TRUE)
}
\arguments{
\item{head.info}{a data.tabl returned by GetHeadlines()}

\item{output}{file to save the results, default setting is to print to scree}

\item{plantuml}{TRUE/FALSE, for plantuml program?}
}
\value{
a string that can be used in plantuml program
}
\description{
Visualise org-mode headings
}
\details{
tree structure of org headlines
}
\author{
Yi Tang
}

