% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/clock.table.R
\name{GetClockTable}
\alias{GetClockTable}
\title{clock.table}
\usage{
GetClockTable(org.file = "~/tmp.org")
}
\arguments{
\item{org.file}{a org file}
}
\value{
a data.table
}
\description{
Parse org file
}
\details{
scan a org file and return the headlines and associated clock entries
}
\author{
Yi Tang
}

